[Dharam] Fixed a bug
What changes have you made?

There is a bug in runtest framework which is preventing the error from
being visual when the error is not on stderr.

Why do we need this change?

Because it was raising an problem where some of the test were failing but
there was no error message.